feat(recipe) : remise en forme des cards pour les recettes
